adding poetry and code ext setup

+++ dotfiles/setup.sh
@@ -1,4 +1,5 @@
 #!/bin/bash
+echo "#####################  Setup log #######################" >> setuplog.txt
 echo "Please enter your email:"
 read user_email
 
@@ -22,10 +23,17 @@ sudo apt upgrade -y && sudo apt autoremove -y
 ####################
 install_package() {
     if ! command -v $1 &> /dev/null; then
-        sudo apt install -y $1
+        if sudo apt install -y $1; then
+            echo "$(date) - $1 installed." >> setuplog.txt
+        else
+            echo "$(date) - $1 is not a valid package." >> setuplog.txt
+        fi
     else
-        echo "$1 is already installed. Updating to the latest version..."
-        sudo apt upgrade -y $1
+        if sudo apt upgrade -y $1; then
+            echo "$(date) - $1 updated." >> setuplog.txt
+        else
+            echo "$(date) - Failed to update $1. Check if it's a valid package." >> setuplog.txt
+        fi
     fi
 }
 
@@ -49,10 +57,12 @@ do
                     sh -c "$(curl -fsSL https://raw.github.com/ohmyzsh/ohmyzsh/master/tools/install.sh)"
                     chsh -s /bin/zsh $user_name
                     sudo apt install fzf -y && sudo apt install zsh-autosuggestions -y
+                    echo "$(date) - $command installed." >> setuplog.txt
                 fi
                 ;;
             "databricks")
                 curl -fsSL https://raw.githubusercontent.com/databricks/setup-cli/main/install.sh | sudo sh
+                echo "$(date) - $command installed." >> setuplog.txt
                 ;;
             "az")
                 curl -sL https://packages.microsoft.com/keys/microsoft.asc | gpg --dearmor | sudo tee /etc/apt/trusted.gpg.d/microsoft.gpg > /dev/null
@@ -60,6 +70,7 @@ do
                 echo "deb [arch=`dpkg --print-architecture` signed-by=/etc/apt/trusted.gpg.d/microsoft.gpg] https://packages.microsoft.com/repos/azure-cli/ $AZ_REPO main" | sudo tee /etc/apt/sources.list.d/azure-cli.list
                 sudo apt update
                 sudo apt install azure-cli -y
+                echo "$(date) - $command installed." >> setuplog.txt
                 ;;
             "git-credential-manager")
                 wget "https://github.com/git-ecosystem/git-credential-manager/releases/download/v2.4.1/gcm-linux_amd64.2.4.1.deb" -O /tmp/gcmcore.deb && sudo dpkg -i /tmp/gcmcore.deb
@@ -67,58 +78,73 @@ do
             "poetry")
                 mkdir -v -p $user_home/.poetry/bin
                 curl -sSL https://install.python-poetry.org | sudo python3 - --prefix=$user_home/.poetry/bin
+                echo "$(date) - $command installed." >> setuplog.txt
                 ;;
         esac
     else
-        echo "$command is already installed."
+        echo "$(date) - $command is already installed." >> setuplog.txt
     fi
 done
 
 ########################################
 # Download the configure-git.sh script #
 ########################################
-configure_git_script_url="https://raw.githubusercontent.com/cltj/dotfiles/master/dotfiles/configure/configure-git.sh"
-configure_git_script="./configure-git.sh"
+read -t 60 -p "Do you want to setup git? (y/n) " answer
 
-if curl -s -o "$configure_git_script" "$configure_git_script_url"; then
-    # Only try to set permissions and run the script if the download succeeded
-    chmod +x "$configure_git_script"
+if [[ $answer =~ ^[Yy]$ ]]
+then
+    configure_git_script_url="https://raw.githubusercontent.com/cltj/dotfiles/master/dotfiles/configure/configure-git.sh"
+    configure_git_script="./configure-git.sh"
 
-    if [ -x "$configure_git_script" ]; then
-        # Only run the child script if it is executable
-        (./configure-git.sh "$user_email" "$user_name" "$user_home") || echo "configure-git.sh script failed"
+    if curl -s -o "$configure_git_script" "$configure_git_script_url"; then
+        # Only try to set permissions and run the script if the download succeeded
+        chmod +x "$configure_git_script"
+
+        if [ -x "$configure_git_script" ]; then
+            # Only run the child script if it is executable
+            (./configure-git.sh "$user_email" "$user_name" "$user_home") || echo "$(date) - configure-git.sh script failed" >> setuplog.txt
+        else
+            echo "$(date) - Failed to set execute permissions on configure-git.sh" >> setuplog.txt
+            echo "$(date) - Please run configure.sh manually" >> setuplog.txt
+        fi
     else
-        echo "Failed to set execute permissions on configure-git.sh"
-        echo "Please run the script manually"
+        echo "$(date) - Failed to download configure-git.sh" >> setuplog.txt
     fi
 else
-    echo "Failed to download configure-git.sh"
+    echo "$(date) - Skipping git setup. Continuing with the script..." >> setuplog.txt
 fi
 
 #############################################
 # Download the configure-dotfiles.sh script #
 #############################################
-configure_dotfiles_script_url="https://raw.githubusercontent.com/cltj/dotfiles/master/dotfiles/configure/configure-dotfiles.sh"
-configure_dotfiles_script="./configure-dotfiles.sh"
+read -t 60 -p "Do you want to setup dotfiles? (y/n) " answer
 
-if curl -s -o "$configure_dotfiles_script" "$configure_dotfiles_script_url"; then
-    # Only try to set permissions and run the script if the download succeeded
-    chmod +x "$configure_dotfiles_script"
+if [[ $answer =~ ^[Yy]$ ]]
+then
+    configure_dotfiles_script_url="https://raw.githubusercontent.com/cltj/dotfiles/master/dotfiles/configure/configure-dotfiles.sh"
+    configure_dotfiles_script="./configure-dotfiles.sh"
 
-    if [ -x "$configure_dotfiles_script" ]; then
-        # Only run the child script if it is executable
-        ./configure-dotfiles.sh || echo "configure-dotfiles.sh script failed"
-        # Source dotfiles.
-        test -e $user_home/.zshrc && source $user_home/.zshrc
-        test -e $user_home/.bashrc && source $user_home/.bashrc
-        test -e $user_home/.config/envman/PATH.env && source $user_home/.config/envman/PATH.env
-        test -e $user_home/.dotfiles/.commonrc && source $user_home/.dotfiles/.commonrc
+    if curl -s -o "$configure_dotfiles_script" "$configure_dotfiles_script_url"; then
+        # Only try to set permissions and run the script if the download succeeded
+        chmod +x "$configure_dotfiles_script"
+
+        if [ -x "$configure_dotfiles_script" ]; then
+            # Only run the child script if it is executable
+            ./configure-dotfiles.sh || echo "$(date) - configure-dotfiles.sh script failed" >> setuplog.txt
+            # Source dotfiles.
+            test -e $user_home/.zshrc && source $user_home/.zshrc
+            test -e $user_home/.bashrc && source $user_home/.bashrc
+            test -e $user_home/.config/envman/PATH.env && source $user_home/.config/envman/PATH.env
+            test -e $user_home/.dotfiles/.commonrc && source $user_home/.dotfiles/.commonrc
+        else
+            echo "$(date) - Failed to set execute permissions on configure-dotfiles.sh" >> setuplog.txt
+            echo "$(date) - Please run configure-dotfiles.sh manually" >> setuplog.txt
+        fi
     else
-        echo "Failed to set execute permissions on configure-dotfiles.sh"
-        echo "Please run the script manually"
+        echo "$(date) - Failed to download configure-dotfiles.sh" >> setuplog.txt
     fi
 else
-    echo "Failed to download configure-dotfiles.sh"
+    echo "$(date) - Skipping dotfiles setup. Continuing with the script..." >> setuplog.txt
 fi
 
 #######################
@@ -132,12 +158,12 @@ then
     echo "Please follow the instructions in your web browser to log in to Azure."
     # Perform an interactive login
     if az login; then
-        echo "Successfully logged in to Azure."
+        echo "$(date) - Successfully logged in to Azure." >> setuplog.txt
     else
-        echo "Failed to log in to Azure. Continuing with the script..."
+        echo "$(date) - Failed to log in to Azure. Continuing with the script..." >> setuplog.txt
     fi
 else
-    echo "Skipping Azure login. Continuing with the script..."
+    echo "$(date) - Skipping Azure login. Continuing with the script..." >> setuplog.txt
 fi
 
 ############################
@@ -151,39 +177,61 @@ then
     # Perform an interactive configuration
     if databricks configure --token; then
         if databricks workspace ls >/dev/null 2>&1; then
-            echo "Successfully verified Databricks CLI configuration."
+            echo "$(date) - Successfully verified Databricks CLI configuration." >> setuplog.txt
         else
-            echo "Failed to verify Databricks CLI configuration. Please check your Databricks host and token."
+            echo "$(date) - Failed to verify Databricks CLI configuration. Please check your Databricks host and token." >> setuplog.txt
         fi
     else
-        echo "Failed to configure Databricks CLI. Continuing with the script..."
+        echo "$(date) - Failed to configure Databricks CLI." >> setuplog.txt
     fi
 else
-    echo "Skipping Databricks CLI configuration. Continuing with the script..."
+    echo "$(date) - Skipping Databricks CLI configuration. Continuing with the script..." >> setuplog.txt
 fi
 
-# Prompt the user for the organization, project, and repository names
-echo "Please enter the Azure DevOps organization name:"
-read organization
-echo "Please enter the Azure DevOps project name:"
-read project
-echo "Please enter the Azure DevOps repository name:"
-read repository
+########################
+# Clone the repository #
+########################
+read -t 60 -p "Do you want to clone repo? (y/n) " answer
 
-# Define the Azure DevOps repository URL
-repo_url="https://dev.azure.com/${organization}/${project}/_git/${repository}"
+if [[ $answer =~ ^[Yy]$ ]]
+then
+    # Prompt the user for the organization, project, and repository names
+    echo "Please enter the Azure DevOps organization name:"
+    read organization
+    echo "Please enter the Azure DevOps project name:"
+    read project
+    echo "Please enter the Azure DevOps repository name:"
+    read repository
 
-# Change to the /mnt/c/dev directory
-cd /mnt/c/dev
+    # Define the Azure DevOps repository URL
+    repo_url="https://dev.azure.com/${organization}/${project}/_git/${repository}"
 
-# Clone the repository
-git clone $repo_url
+    # Change to the /mnt/c/dev directory
+    cd /mnt/c/dev
 
-# Verify that the repository was cloned
-if [ $? -eq 0 ]; then
-    echo "Successfully cloned the repository."
-    cd $repository
-    code .
+    # Clone the repository
+    git clone $repo_url
+
+    # Verify that the repository was cloned
+    if [ $? -eq 0 ]; then
+        echo "$(date) - Successfully cloned the repository." >> setuplog.txt
+        cd $repository
+        code .
+        # Check if Visual Studio Code is running
+        if pgrep -x "code" > /dev/null
+        then
+            echo "Visual Studio Code is running." >> setuplog.txt
+            ./configure-vscode-extentions.sh || echo "Failed to run configure-vscode-extentions.sh" >> setuplog.txt
+            ./configure-poetry.sh $repository $user_home || echo "Failed to run configure-poetry.sh" >> setuplog.txt
+            source $user_home/.bashrc
+        else
+            echo "$(date) - Visual Studio Code failed to start." >> setuplog.txt
+        fi
+    else
+        echo "$(date) - Failed to clone the repository." >> setuplog.txt
+    fi
 else
-    echo "Failed to clone the repository."
+    echo "$(date) - Skipping repo cloning. Continuing with the script..." >> setuplog.txt
 fi
+
+echo "#####################  setup.sh done! #######################" >> setuplog.txt
